refactor: Rename battery_harvester to batteries
