+++ src/app/data_harvester/batteries.rs
